Rename DialogComponent to BaseDialog
DialogComponent does so little, that I'm going to build a richer
Dialog component on top of that which should be the standard
component to use unless you need something super custom. As such, I've
renamed DialogComponent to BaseDialog to reflect it's primordial state.

I also added a `useDeprecatedGlobalStyles` flag which adds the various
css classes that trigger the globally defined css styles which we would
like to get away from.
